refactor: moving dataprovider package to infra
