+++ src/main/java/com/Reviewz/core/authentication/usecase/AuthenticationService.java
@@ -6,7 +6,7 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-import com.Reviewz.dataprovider.database.UserRepository;
+import com.Reviewz.infra.dataprovider.database.UserRepository;
 
 @Service
 public class AuthenticationService implements UserDetailsService{
